+++ examples/simple/token.ts
@@ -22,41 +22,17 @@
  * SOFTWARE.
  */
 
-import type { ExpressContext } from 'apollo-server-express';
-import type { Context } from './Context';
-import { TokenPayload, verify } from './token';
-import { AuthenticationError } from '../../src';
+import jwt from 'jsonwebtoken';
+import type { User } from './User';
 
-export async function contextHelper({ req }: ExpressContext): Promise<Context> {
-  let user: TokenPayload | undefined;
-  const authorizationHeader =
-    req.headers && 'Authorization' in req.headers
-      ? 'Authorization'
-      : 'authorization';
+export type TokenPayload = Omit<User, 'secret'>;
 
-  if (req.headers && req.headers[authorizationHeader]) {
-    const parts = (req.headers[authorizationHeader] as string).split(' ');
+const SECRET = 'secret';
 
-    if (parts.length === 2) {
-      const scheme = parts[0];
-      const credentials = parts[1];
+export function sign(payload: TokenPayload) {
+  return jwt.sign(payload, SECRET);
+}
 
-      if (/^Bearer$/i.test(scheme)) {
-        const token = credentials;
-
-        try {
-          const decodedToken = verify(token);
-          user = decodedToken;
-        } catch (error) {
-          throw new AuthenticationError();
-        }
-      }
-    } else {
-      throw new AuthenticationError(
-        "Token format is 'Authorization: Bearer [token]'"
-      );
-    }
-  }
-
-  return { user };
+export function verify(token: string) {
+  return jwt.verify(token, SECRET) as TokenPayload;
 }