+++ examples/typegraphql/User.ts
@@ -22,41 +22,24 @@
  * SOFTWARE.
  */
 
-import type { ExpressContext } from 'apollo-server-express';
-import type { Context } from './Context';
-import { TokenPayload, verify } from './token';
-import { AuthenticationError } from '../../src';
+import { GraphQLInt } from 'graphql';
+import { ObjectType, Field } from 'type-graphql';
+import { UserRoles } from './UserRoles';
+import { UserPermissions } from './UserPermissions';
+import { Auth } from './Auth';
 
-export async function contextHelper({ req }: ExpressContext): Promise<Context> {
-  let user: TokenPayload | undefined;
-  const authorizationHeader =
-    req.headers && 'Authorization' in req.headers
-      ? 'Authorization'
-      : 'authorization';
+@ObjectType()
+export class User {
+  @Field(() => GraphQLInt)
+  id!: number;
 
-  if (req.headers && req.headers[authorizationHeader]) {
-    const parts = (req.headers[authorizationHeader] as string).split(' ');
+  @Field()
+  @Auth({ roles: [UserRoles.ADMIN] })
+  secret!: boolean;
 
-    if (parts.length === 2) {
-      const scheme = parts[0];
-      const credentials = parts[1];
+  @Field(() => [UserRoles])
+  roles!: UserRoles[];
 
-      if (/^Bearer$/i.test(scheme)) {
-        const token = credentials;
-
-        try {
-          const decodedToken = verify(token);
-          user = decodedToken;
-        } catch (error) {
-          throw new AuthenticationError();
-        }
-      }
-    } else {
-      throw new AuthenticationError(
-        "Token format is 'Authorization: Bearer [token]'"
-      );
-    }
-  }
-
-  return { user };
+  @Field(() => [UserPermissions])
+  permissions!: UserPermissions[];
 }