+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Changelog and release notes
 
+## v0.2.0
+
+### Features
+
+- Accept generics `TRole` and `TPermission` along with `TContext`
+
+- `TypeGraphQL` integration example
+
 ## v0.1.1
 
-### Fix Typescript paths
+### Fixes
+
+- Typescript paths
 
 ## v0.1.0